+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 * Fix visit and event query by properly merging the searches
   - https://activerecord-hackery.github.io/ransack/going-further/merging-searches/
   - this requires queries to specify the full table name 
-  
+* Rewrite bounce rate query
+
 ### Changed
 
 * Changed the configuration file to properly reflect method names